chore: Move walker protocol
